Migrate to new monitor-type
